SignUpActivity refactored.
