+++ app/src/main/java/com/alkemy/ongandroid/core/AppConstants.kt
@@ -1,4 +1,4 @@
-package com.alkemy.ongandroid.view
+package com.alkemy.ongandroid.core
 
 object AppConstants {
