+++ app/src/main/java/com/alkemy/ongandroid/viewmodel/SignUpViewModel.kt
@@ -6,10 +6,10 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.alkemy.ongandroid.businesslogic.managers.LocalDataManager
+import com.alkemy.ongandroid.core.AppConstants
 import com.alkemy.ongandroid.model.NewUserResponse
 import com.alkemy.ongandroid.model.User
 import com.alkemy.ongandroid.model.UserRepository
-import com.alkemy.ongandroid.view.AppConstants
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -50,28 +50,27 @@ class SignUpViewModel @Inject constructor(
     }
 
     private val _isButtonEnabled = MutableLiveData(false)
-    val isButtonSaveEnabled : LiveData<Boolean> = _isButtonEnabled
+    val isButtonSaveEnabled: LiveData<Boolean> = _isButtonEnabled
 
-    fun validateFields(username:String, email:String, password: String, confirmPassword:String) {
+    fun validateFields(username: String, email: String, password: String, confirmPassword: String) {
 
-        val fieldsEmpty: Boolean = username.isEmpty() || email.isEmpty() || password.isEmpty() || confirmPassword.isEmpty()
+        val fieldsEmpty: Boolean =
+            username.isEmpty() || email.isEmpty() || password.isEmpty() || confirmPassword.isEmpty()
         val emailFormat: Boolean = Patterns.EMAIL_ADDRESS.matcher(email).matches()
-        val passwordsFormat: Boolean = isValidPasswordFormat(password) && isValidPasswordFormat(confirmPassword)
-
+        val passwordsFormat: Boolean =
+            isValidPasswordFormat(password) && isValidPasswordFormat(confirmPassword)
 
         _isButtonEnabled.value = !fieldsEmpty && emailFormat && passwordsFormat
     }
 
     private val _arePasswordsTheSame = MutableLiveData(false)
-    val arePasswordsTheSame : LiveData<Boolean> = _isButtonEnabled
+    val arePasswordsTheSame: LiveData<Boolean> = _isButtonEnabled
 
     fun comparePasswords(password: String, confirmPassword: String) {
         _arePasswordsTheSame.value = (password == confirmPassword)
     }
 
-
     private fun isValidPasswordFormat(password: String): Boolean {
         return Pattern.matches(AppConstants.PASSWORD_REGEX, password)
     }
-
 }
